# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ETK API",
        "contact": {
            "name": "fbeck",
            "email": "loordy@live.ru"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        },
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        }
    ],
    "paths": {
...
        "/structures/": {
            "post": {
                "tags": [
                    "Department"
                ],
                "operationId": "/structures/",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StructuresStore"
                },
                "responses": {
                    "200": {
                        "description": "Department saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Structures"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Contract retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to create this structure"
                    },
                    "404": {
                        "description": "Parent structure not found"
                    },
                    "424": {
                        "description": "SMTH wrong with dates"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
        "/structure/{id}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "operationId": "/structure/",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Structures"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Contract retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to show this structure"
                    },
                    "404": {
                        "description": "Department not found"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Department"
                ],
                "operationId": "/structure/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/StructuresUpdate"
                },
                "responses": {
                    "200": {
                        "description": "Department updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Structures"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Contract updated successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to update this structure"
                    },
                    "404": {
                        "description": "Parent structure not found"
                    },
                    "424": {
                        "description": "SMTH wrong with dates"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Department"
                ],
                "operationId": "/structure/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/StructuresDestroy"
                },
                "responses": {
                    "200": {
                        "description": "Department stoped successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Department stoped successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to update this structure"
                    },
                    "404": {
                        "description": "Parent structure not found"
                    },
                    "424": {
                        "description": "SMTH wrong with dates"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
        "/structures/{tin}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "operationId": "/structures/{tin}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Structures"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Contract retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
        "/company/{tin}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "operationId": "/company/{tin}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Structures"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Contract retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Structures": {
                "title": "Структура организации",
                "description": "Структура организации",
                "required": [
                    "id_structure",
                    "name_structure"
                ],
                "properties": {
                    "id_structure": {
                        "title": "Уникальный идентификатор структуры",
                        "description": "Уникальный идентификатор структуры",
                        "format": "int64"
                    },
                    "parent_id_structure": {
                        "title": "Родительская запись стр.орг.",
                        "description": "Родительская запись стр.орг.",
                        "format": "int64"
                    },
                    "name_structure": {
                        "title": "Наименование структуры",
                        "description": "Наименование структуры",
                        "format": "string"
                    },
                    "tin_company_structure": {
                        "title": "ИНН ЮЛ в структуре организации",
                        "description": "ИНН ЮЛ в структуре организации",
                        "format": "string"
                    },
                    "oked_structure": {
                        "title": "Код ОКЭД",
                        "description": "Код ОКЭД",
                        "format": "string"
                    },
                    "region_structure": {
                        "title": "СОАТО структуры",
                        "description": "СОАТО структуры",
                        "format": "string"
                    },
                    "region4_structure": {
                        "title": "СОАТО структуры 4 знака",
                        "description": "СОАТО структуры 4 знака",
                        "format": "int64"
                    },
                    "root_id_structure": {
                        "title": "Ссылка на корневую запись структуры",
                        "description": "Ссылка на корневую запись структуры",
                        "format": "int64"
                    },
                    "data_structure": {
                        "title": "",
                        "description": "",
                        "format": "array"
                    },
                    "date_end_structure": {
                        "title": "Дата конца",
                        "description": "",
                        "format": "date"
                    },
                    "date_start_structure": {
                        "title": "Дата начала",
                        "description": "",
                        "format": "date"
                    }
                },
                "type": "object",
                "example": {
                    "id_structure": 2044,
                    "parent_id_structure": 5,
                    "name_structure": "Отдел чего то там2",
                    "tin_company_structure": "123123123",
                    "oked_structure": null,
                    "region_structure": null,
                    "region4_structure": null,
                    "root_id_structure": "6",
                    "data_structure": null,
                    "date_end_structure": null,
                    "date_start_structure": "2017-12-12"
                }
            },
...
        },
        "requestBodies": {
...
            "StructuresStore": {
                "description": "Department object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Структура организации",
                                "description": "Структура организации",
                                "required": [
                                    "parent_id_structure",
                                    "name_structure",
                                    "date_start_structure"
                                ],
                                "example": {
                                    "parent_id_structure": 5,
                                    "name_structure": "Отдел #2",
                                    "date_end_structure": "2018-12-12",
                                    "date_start_structure": "2017-12-12"
                                }
                            }
                        }
                    }
                }
            },
            "StructuresUpdate": {
                "description": "Department object that needs to be updated in the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Структура организации",
                                "description": "Структура организации",
                                "required": [
                                    "parent_id_structure",
                                    "name_structure",
                                    "date_start_structure"
                                ],
                                "example": {
                                    "parent_id_structure": 5,
                                    "name_structure": "Отдел #2",
                                    "date_end_structure": "2018-12-12",
                                    "date_start_structure": "2017-12-12"
                                }
                            }
                        }
                    }
                }
            },
            "StructuresDestroy": {
                "description": "Department object that needs to be destroyed in the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Структура организации",
                                "description": "Структура организации",
                                "required": [],
                                "example": []
                            }
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "etk_auth": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}